requirements.txt
discord.py
tweepy==3.10.0
asyncio
Pillow >= 8.0
requests_oauthlib
urllib3
datetime
pytz==2022.1
